Add dynamic data to frontend design - Refactor modal styles into separate file
