+++ client/src/components/styledComponents/ModalStyles.jsx
@@ -1,0 +1,17 @@
+const MODAL_STYLE = {
+  position: 'fixed',
+  padding: '.2% 3%',
+  zIndex: 1000
+};
+
+const OVERLAY_STYLE = {
+  position: 'fixed',
+  top: 0,
+  left: 0,
+  right: 0,
+  bottom: 0,
+  backgroundColor: 'rgba(0, 0, 0, .89)',
+  zIndex: 1000
+};
+
+export { MODAL_STYLE, OVERLAY_STYLE }
